do moduling
