+++ python/code_challenges/tree_fizz_buzz/tree_fizz_buzz/queue.py
@@ -1,0 +1,42 @@
+
+class Queue:
+    def __init__(self):
+        self.front=None
+        self.rear=None
+    def enqueue(self,value):
+      
+        if self.is_empty():
+            self.front=value
+            self.rear=value
+        self.rear.next=value
+        self.rear=value
+
+    def dequeue(self):
+        if self.is_empty():
+            raise Exception("empty equeue")
+        if self.front==self.rear:
+            temp=self.front
+            self.front=None
+            self.rear=None
+            return temp.value
+        else:
+            temp=self.front
+            self.front=self.front.next
+            temp.next=None
+            return temp.value
+
+    def peek(self):
+       if self.is_empty():
+           raise Exception("empty equeue")
+       return self.front.value
+
+
+    def is_empty(self):
+     return not self.front
+
+    def __len__(self):
+        counter=0
+        while self.front:
+            counter +=1
+            self.dequeue()
+        return counter
