+++ python/code_challenges/tree_fizz_buzz/tree_fizz_buzz/tree_fizz_buzz.py
@@ -1,52 +1,9 @@
 import copy
-
+from tree_fizz_buzz.queue import *
 class Node:
     def __init__(self,value):
         self.value = value
         self.children = []
-
-
-class Queue:
-    def __init__(self):
-        self.front=None
-        self.rear=None
-    def enqueue(self,value):
-      
-        if self.is_empty():
-            self.front=value
-            self.rear=value
-        self.rear.next=value
-        self.rear=value
-
-    def dequeue(self):
-        if self.is_empty():
-            raise Exception("empty equeue")
-        if self.front==self.rear:
-            temp=self.front
-            self.front=None
-            self.rear=None
-            return temp.value
-        else:
-            temp=self.front
-            self.front=self.front.next
-            temp.next=None
-            return temp.value
-
-    def peek(self):
-       if self.is_empty():
-           raise Exception("empty equeue")
-       return self.front.value
-
-
-    def is_empty(self):
-     return not self.front
-
-    def __len__(self):
-        counter=0
-        while self.front:
-            counter +=1
-            self.dequeue()
-        return counter
 
 
 class k_ary_tree:
@@ -105,7 +62,7 @@ if __name__ == "__main__":
     new_tree =  fizz_buzz_tree(tree)
 
     # print(new_tree.root.children[1].children[1].value)
-    print(new_tree.root.value)
+    print(new_tree.root.children[1].value)
 
     # print(dir([]))
     # a = new_tree.__init__()
